+++ app/src/main/java/com/android04/godfisherman/ui/main/MainActivity.kt
@@ -2,11 +2,17 @@ package com.android04.godfisherman.ui.main
 
 import android.content.Intent
 import android.annotation.SuppressLint
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.IntentFilter
 import android.content.pm.ActivityInfo
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import androidx.activity.viewModels
 import androidx.constraintlayout.motion.widget.MotionLayout
+import androidx.core.app.NotificationManagerCompat
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.android04.godfisherman.R
 import com.android04.godfisherman.databinding.ActivityMainBinding
@@ -19,39 +25,76 @@ import com.android04.godfisherman.ui.stopwatch.StopwatchInfoFragment
 import com.android04.godfisherman.ui.stopwatch.TestStopwatchFragment
 import com.android04.godfisherman.utils.BindingAdapter
 import com.android04.godfisherman.utils.StopwatchNotification
+import com.android04.godfisherman.utils.StopwatchService
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.activity_main) {
 
     override val viewModel: MainViewModel by viewModels()
-  
+
+    companion object {
+        var isStopwatchServiceRunning = false
+    }
+    private lateinit var serviceIntent: Intent
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setObserver()
         viewModel.checkConnectivity()
-
+        binding.viewModel = viewModel
         setOrientation()
         
         StopwatchNotification.createChannel(this)
         initBottomNavigation()
         initMotionListener()
-
+//        Log.d("StopWatch", "isFromService : ${MainViewModel.isFromService}")
+//        if (MainViewModel.isFromService) {
+//            viewModel.stopwatchOnFlag.value = true
+//        }
         viewModel.stopwatchOnFlag.observe(this) { flag ->
             if (flag) {
+                Log.d("MotionLayout", "옵저브 실행")
                 // TODO 수정해야함
                 // 라이브데이터 바인딩이 제대로 되지 않아서 일단 임시 방편으로 바인딩 어댑터를 직접 사용
                 BindingAdapter.setVisibilityOnMotion(binding.clContainerStopwatch, flag)
                 changeFragment(R.id.fl_stopwatch_big, TestStopwatchFragment())
                 binding.container.setTransition(R.id.transition)
                 binding.container.transitionToState(R.id.end)
+            } else {
+                BindingAdapter.setVisibilityOnMotion(binding.clContainerStopwatch, flag)
             }
         }
 
         binding.ivClose.setOnClickListener {
-            binding.container.transitionToState(R.id.end_close)
-            viewModel.stopwatchOnFlag.value = false
+            if (viewModel.isStopwatchStarted.value == true) {
+                binding.container.transitionToState(R.id.end)
+            } else {
+                binding.container.transitionToState(R.id.end_close)
+                viewModel.stopwatchOnFlag.value = false
+            }
         }
+
+
+        serviceIntent = Intent(this, StopwatchService::class.java)
+        registerReceiver(receiveTime, IntentFilter(StopwatchService.SERVICE_DESTROYED))
+        val flag = intent.getBooleanExtra(StopwatchService.FROM_SERVICE, false)
+        MainViewModel.isFromService = flag
+        Log.d("StopWatch", "받은 데이터 : $flag")
+        if (flag) {
+            Log.d("StopWatch", "받은 메소드 실행")
+            viewModel.stopwatchOnFlag.value = true
+        }
+
+        viewModel.isAfterUpload.observe(this) {
+            if (it){
+                binding.container.transitionToState(R.id.end_close)
+                viewModel.stopwatchOnFlag.value = false
+                viewModel.setIsAfterUploadFalse()
+            }
+        }
+
     }
 
     private fun initBottomNavigation() {
@@ -68,6 +111,7 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.a
                     true
                 }
                 R.id.navigation_camera -> {
+                    viewModel.isServiceRequestWithOutCamera = false
                     startActivity(Intent(this, CameraActivity::class.java))
                     viewModel.beforeMenuItemId = R.id.navigation_camera
                     true
@@ -105,6 +149,7 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.a
 
             override fun onTransitionCompleted(motionLayout: MotionLayout?, currentId: Int) {
 
+
                 if (viewModel.isFromInfoFragment) {
                     supportFragmentManager.beginTransaction().replace(R.id.fl_fragment_container, HomeFragment()).commit()
                     viewModel.isFromInfoFragment = false
@@ -114,6 +159,7 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.a
                     binding.navView.menu.findItem(viewModel.beforeMenuItemId).isChecked = true
                     viewModel.isFromStopwatchFragment = false
                 }
+                MainViewModel.isFromService = false
                 viewModel.isOpened = currentId == R.id.end
             }
 
@@ -139,4 +185,45 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.a
         viewModel.isNetworkConnected.observe(this) {
         }
     }
+
+    override fun onStop() {
+        super.onStop()
+        if (viewModel.isServiceRequestWithOutCamera) {
+            Log.d("StopWatch", "서비스 실행 요청")
+            passStopwatchToService(viewModel.time)
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        viewModel.isServiceRequestWithOutCamera = true
+        Log.d("serviceRunning", "${isStopwatchServiceRunning}")
+        if (isStopwatchServiceRunning) {
+            Log.d("serviceRunning", "서비스 종료 실행")
+            stopService(serviceIntent)
+            isStopwatchServiceRunning = false
+        }
+    }
+
+    private fun passStopwatchToService(time: Double) {
+        if (viewModel.isStopwatchStarted.value == true) {
+            viewModel.passStopwatchToService()
+            serviceIntent.putExtra(StopwatchService.TIME_EXTRA, time)
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                startForegroundService(serviceIntent)
+            } else {
+                startService(serviceIntent)
+            }
+        }
+    }
+
+    private val receiveTime: BroadcastReceiver = object : BroadcastReceiver() {
+        override fun onReceive(context: Context, receiveIntent: Intent) {
+            val receivedTime = receiveIntent.getDoubleExtra(StopwatchService.SERVICE_DESTROYED, 0.0)
+            Log.d("receiveFromService", "리시브 실행 : $receivedTime")
+            viewModel.passedTimeFromService(receivedTime)
+            NotificationManagerCompat.from(context).cancel(StopwatchService.NOTIFICATION_ID)
+        }
+    }
+
 }
