+++ app/src/main/java/com/android04/godfisherman/ui/stopwatch/StopwatchInfoFragment.kt
@@ -1,6 +1,5 @@
 package com.android04.godfisherman.ui.stopwatch
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
@@ -11,23 +10,14 @@ import com.android04.godfisherman.ui.main.MainActivity
 
 class StopwatchInfoFragment : BaseFragment<FragmentStopwatchInfoBinding, StopwatchInfoViewModel>(R.layout.fragment_stopwatch_info) {
     override val viewModel: StopwatchInfoViewModel by viewModels()
-    lateinit var intent: Intent
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        intent = Intent(requireContext(), StopwatchActivity::class.java)
-        skipInfo()
-//        binding.btnStart.setOnClickListener { startActivity(intent) }
+
         binding.btnStart.setOnClickListener {
             val activity = requireActivity() as MainActivity
             activity.viewModel.stopwatchOnFlag.value = true
             activity.viewModel.isFromInfoFragment = true
-        }
-    }
-
-    private fun skipInfo(){
-        if (StopwatchActivity.isStopwatchServiceRunning) {
-            startActivity(intent)
         }
     }
 }
