+++ app/src/main/java/com/android04/godfisherman/utils/StopwatchService.kt
@@ -9,11 +9,9 @@ import android.util.Log
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import com.android04.godfisherman.R
-import com.android04.godfisherman.ui.stopwatch.StopwatchActivity
-import kotlinx.coroutines.delay
-import java.lang.Thread.sleep
+import com.android04.godfisherman.ui.main.MainActivity
+import com.android04.godfisherman.ui.stopwatch.TestStopwatchFragment
 import java.util.*
-import kotlin.concurrent.thread
 
 class StopwatchService:
     Service() {
@@ -21,6 +19,7 @@ class StopwatchService:
     companion object {
         const val TIME_EXTRA = "timeExtra"
         const val TIME_SAVED = "timeSaved"
+        const val FROM_SERVICE = "fromService"
         const val SERVICE_DESTROYED = "serviceDestroyed"
         const val STOPWATCH_ENTER = "timeEnter"
         const val NOTIFICATION_ID = 10
@@ -32,12 +31,13 @@ class StopwatchService:
     override fun onBind(p0: Intent?): IBinder? = null
 
     override fun onStartCommand(passedIntent: Intent, flags: Int, startId: Int): Int {
-        StopwatchActivity.isStopwatchServiceRunning = true
-        val intent = Intent(this, StopwatchActivity::class.java)
+        Log.d("StopWatch", "서비스 실행")
+        MainActivity.isStopwatchServiceRunning = true
+        val intent = Intent(this, MainActivity::class.java)
         intent.action = STOPWATCH_ENTER
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or
                 Intent.FLAG_ACTIVITY_CLEAR_TASK
-        intent.putExtra(TIME_SAVED, saveTime)
+        intent.putExtra(FROM_SERVICE, true)
         val pendingIntent = PendingIntent
             .getActivity(this, 0, intent, PendingIntent.FLAG_IMMUTABLE)
